+++ components/snackbar/index.js
@@ -1,24 +1,13 @@
-import React from 'react';
+import React from "react";
 import Portal from "../portal";
 import SnackBar from "./styled-snackbar";
 
-const Snackbar = (props) => {    
-    
-    return (
-        <Portal >
-            <SnackBar {...props}            
-            >
-                
-                    {props.children}
-               
-            </SnackBar>
-                
-            
-        </Portal>
-        
-    ); 
+const Snackbar = (props) => {
+  return (
+    <Portal>
+      <SnackBar {...props}>{props.children}</SnackBar>
+    </Portal>
+  );
 };
-
-  
 
 export default Snackbar;
