+++ components/snackbar/styled-snackbar.js
@@ -1,18 +1,12 @@
-import { styled } from "linaria/react";
-
-
+import styled from "styled-components";
 
 const SnackBar = styled.div`
-bottom: 12px;
-min-height: 40px;
-position: fixed;
-width: 100%;
-z-index: 10;
-display: block;
-
-
+  bottom: 12px;
+  min-height: 40px;
+  position: fixed;
+  width: 100%;
+  z-index: 10;
+  display: block;
 `;
-
-
 
 export default SnackBar;
