+++ src/locales/en/translations.json
@@ -64,7 +64,7 @@
   "AuthDocsYourEmail": "Your e-mail*",
   "AuthDocsAccessYourDocs": "Access your documents from anywhere.",
   "CookieMessButton": "Got it!",
-  "CookieMessText": "This website uses cookies. By continuing to browse the website you agree to our {0}privacy policy{1}",
+  "CookieMessText": "This website uses cookies. By continuing to browse the website you agree to our <1>privacy policy</1>",
   "CreatePersonalButton": "Create ONLYOFFICE account",
   "Email": "E-mail",
   "PersonalLogin": "Account Login",
